requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
tenacity
torch
fastmcp
arxiv
pymupdf
numpy